requirements.txt
fastapi
uvicorn[standard]
litellm
python-dotenv~=1.0.0
httpx
rich
pandas
requests
pydantic
SQLAlchemy~=2.0.0
Alembic~=1.13.0
numpy
scikit-learn
judgy
rank-bm25
tqdm
pydantic
typing-extensions
matplotlib
seaborn
plotly
